requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-debug-toolbar==1.11
django-templated-mail==1.1.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
djoser==1.5.0
Faker==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
